Size.c

+++ 0x00-hello_world/6-size.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 /**
  * main - A program that prints the size of various computer types
- * Return 0 (Success)
+ * Return: 0 (Success)
  */
 int main(void)
 {
@@ -10,10 +10,10 @@ int main(void)
 	long int c;
 	long long int d;
 	float f;
-printf("Size of a char: %lu byte(S)\n", (unsigned long)sizeof(a));
-printf("Size of an int: %lu byte(S)\n", (unsigned long)sizeof(b));
-printf("Size of a long int: %lu byte(S)\n", (unsigned long)sizeof(c));
-printf("Size of a long long int: %lu byte(S)\n", (unsigned long)sizeof(d));
-printf("Size of a float: %lu byte(S)\n", (unsigned long)sizeof(f));
+printf("Size of a char: %d byte(S)\n", sizeof(a));
+printf("Size of an int: %d byte(S)\n", sizeof(b));
+printf("Size of a long int: %d byte(S)\n", sizeof(c));
+printf("Size of a long long int: %d byte(S)\n", sizeof(d));
+printf("Size of a float: %d byte(S)\n", sizeof(f));
 return (0);
 }
